pyproject.toml
[project]
name = "acute_resp_agent_job"
version = "0.0.1"
authors = [{ name = "mt3franca@gmail.com" }]
requires-python = ">= 3.11"
dependencies = [
    "pandas>=2.2.3",
    "databricks-feature-engineering>=0.12.1",
    "dotenv>=0.9.9",
    "databricks-langchain>=0.7.0",
    "langchain-core>=0.3.74",
    "langgraph>=0.6.6",
    "toml>=0.10.2",
    "langchain-tavily>=0.2.11",
    "tavily-python>=0.7.11",
    "databricks-sql-connector>=4.1.2",
    "sqlalchemy>=2.0.43",
    "databricks-sqlalchemy>=2.0.7",
    "pdfplumber>=0.11.7",
    "jinja2>=3.1.6",
    "html-to-markdown>=2.3.0",
    "render-html",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = [
    "tests",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/acute_resp_agent_job"]

[project.scripts]
main = "acute_resp_agent_job.main:main"
